+++ data/output copy.json
@@ -6,7 +6,7 @@
     "961193871268794458": [
         961186709897351178
     ],
-    "961131970442694676": [],
+    
     "738354066127847505": [
         961131970442694676
     ],
