+++ data/rank.json
@@ -1,8 +1,8 @@
 {
     "534243081135063041": {
-        "win": 17,
+        "win": 23,
         "lose": 0,
-        "\u6392\u540d": 11
+        "\u6392\u540d": 10
     },
     "450994523443560458": {
         "win": 19,
@@ -247,7 +247,7 @@
     "922498560795238430": {
         "win": 21,
         "lose": 0,
-        "\u6392\u540d": 10
+        "\u6392\u540d": 11
     },
     "533244011771723798": {
         "win": 32,
@@ -270,7 +270,7 @@
         "\u6392\u540d": 11
     },
     "914390138623238175": {
-        "win": 1,
+        "win": 3,
         "lose": 0,
         "\u6392\u540d": 11
     },
@@ -312,8 +312,8 @@
             "win": 32
         },
         "10": {
-            "id": "922498560795238430",
-            "win": 21
+            "id": "534243081135063041",
+            "win": 23
         }
     },
     "741620718575288380": {
